refactor(api): extract dashboard helpers + getProfileSessions to focused service modules
Move pure helpers and learner session logic off the dashboard service so
the progress route no longer depends on a parent-facing aggregator.

- progress-helpers.ts: calculateTrend, calculateRetentionTrend,
  calculateGuidedRatio, buildProgressGuidance
- session/session-crud.ts: getProfileSessions + supporting helpers
- session/session-analytics.ts (new): countGuidedMetrics,
  countGuidedMetricsBatch
- routes/progress.ts: import getProfileSessions from session-crud
- dashboard.ts: re-export shims preserved so internal consumers compile
- sortSubjectsByActivityPriority intentionally kept in dashboard.ts
  (still used; covered by tests)

Validation: api typecheck clean. 19/19 dashboard.helpers tests pass with
zero modifications. 11/11 dashboard.integration tests pass.
10/10 progress-routes.integration tests pass including new
parent-proxy session-list scoping case.

Plan: docs/superpowers/plans/2026-05-13-surface-ownership-boundaries.md
(PR 2 of 10)
